serverless/audio-transcription/requirements.txt
google-cloud-storage==2.5.0
torch
torchaudio
transformers==4.28.0